reorg langs and fixtures
