+++ tests/casper/conftest.py
@@ -4,35 +4,17 @@ from state_languages.blockchain_test_lang import BlockchainTestLang
 from state_languages.integer_test_lang import IntegerTestLang
 from state_languages.binary_test_lang import BinaryTestLang
 
-@pytest.fixture
-def binary_lang(report, test_weight):
-    return BinaryTestLang(test_weight, report)
 
 @pytest.fixture
 def blockchain_lang(report, test_weight):
     return BlockchainTestLang(test_weight, report)
 
-@pytest.fixture
-def integer_lang(report, test_weight):
-    return IntegerTestLang(test_weight, report)
 
-
-@pytest.fixture
-def binary_lang_runner(report):
-    def runner(weights, test_string):
-        BinaryTestLang(weights, report).parse(test_string)
-    return runner
 
 @pytest.fixture
 def blockchain_lang_runner(report):
     def runner(weights, test_string):
         BlockchainTestLang(weights, report).parse(test_string)
-    return runner
-
-@pytest.fixture
-def integer_lang_runner(report):
-    def runner(weights, test_string):
-        IntegerTestLang(weights, report).parse(test_string)
     return runner
 
 
@@ -42,14 +24,9 @@ def binary_lang_creator(report):
         return BinaryTestLang(weights, report)
     return creator
 
+
 @pytest.fixture
 def blockchain_lang_creator(report):
     def creator(weights):
         return BlockchainTestLang(weights, report)
-    return creator
-
-@pytest.fixture
-def integer_lang_creator(report):
-    def creator(weights):
-        return IntegerTestLang(weights, report)
     return creator
