+++ tests/casper/protocols/binary/conftest.py
@@ -1,7 +1,27 @@
 import random
 import pytest
 
+from state_languages.binary_test_lang import BinaryTestLang
 from casper.protocols.binary.binary_protocol import BinaryProtocol
+
+@pytest.fixture
+def binary_lang(report, test_weight):
+    return BinaryTestLang(test_weight, report)
+
+
+@pytest.fixture
+def binary_lang_runner(report):
+    def runner(weights, test_string):
+        BinaryTestLang(weights, report).parse(test_string)
+    return runner
+
+
+@pytest.fixture
+def binary_lang_creator(report):
+    def creator(weights):
+        return BinaryTestLang(weights, report)
+    return creator
+
 
 @pytest.fixture
 def binary_validator_set(generate_validator_set):
