+++ tests/casper/protocols/blockchain/test_blockchain_view.py
@@ -1,18 +1,11 @@
 """The BlockchainView testing module..."""
 import pytest
 
-from casper.protocols.blockchain.blockchain_protocol import BlockchainProtocol
-from state_languages.blockchain_test_lang import BlockchainTestLang
-
-
-@pytest.mark.skip(reason="test not yet implemented")
-def test_add_justified_message():
+@pytest.mark.skip(reason="test language not written")
+def test_update_safe_estimates():
     pass
 
-@pytest.mark.skip(reason="test not yet implemented")
-def test_dont_add_non_justified_message():
-    pass
 
-@pytest.mark.skip(reason="test not yet implemented")
-def test_resolve_non_justified_message_when_justification_arrives():
+@pytest.mark.skip(reason="test language not written")
+def test_update_protocol_specific_view():
     pass
