+++ tests/casper/protocols/integer/conftest.py
@@ -1,7 +1,27 @@
 import random
 import pytest
 
+from state_languages.integer_test_lang import IntegerTestLang
 from casper.protocols.integer.integer_protocol import IntegerProtocol
+
+@pytest.fixture
+def integer_lang(report, test_weight):
+    return IntegerTestLang(test_weight, report)
+
+
+@pytest.fixture
+def integer_lang_runner(report):
+    def runner(weights, test_string):
+        IntegerTestLang(weights, report).parse(test_string)
+    return runner
+
+
+@pytest.fixture
+def integer_lang_creator(report):
+    def creator(weights):
+        return IntegerTestLang(weights, report)
+    return creator
+
 
 @pytest.fixture
 def integer_validator_set(generate_validator_set):
